added svg export

+++ src/main/java/fr/univ_amu/l3mi/drawing_app/view/configuration/CanvasDimensions.java
@@ -1,4 +1,4 @@
 package fr.univ_amu.l3mi.drawing_app.view.configuration;
 
-public record CanvasDimensions(int width, int height) {
+public record CanvasDimensions(double width, double height) {
 }
